requirements.txt
Django==4.2.9
djangorestframework==3.14.0
drf-yasg==1.21.7
python-dateutil==2.8.2
setuptools>=65.0.0
django-cors-headers==4.9.0
